static shape padding style to batch inside,new unit test method, real classnum shape
