+++ inference-engine/tests/functional/shared_test_classes/include/shared_test_classes/single_layer/multiclass_nms.hpp
@@ -51,7 +51,9 @@ protected:
     void SetUp() override;
 
 private:
-    size_t numOfSelectedBoxes;
+    size_t numBatches, numBoxes, numClasses;
+    size_t maxOutputBoxesPerClass;
+    size_t maxOutputBoxesPerBatch;
 };
 
 }  // namespace LayerTestsDefinitions
